requirements.txt
fastapi
pydantic
sqlalchemy
python-multipart
python-jose
bcrypt
cryptography
user-agents
uvicorn